Practice with Python from 12/1/15
Created functions in Time.py to calculate hours, minutes, and seconds,
based on a number of seconds.
Created functions in Money.py to calculate the number of quarters,
dimes, nickels, and pennies, based on a number of pennies.
Implemented the functions from Money.py and Time.py in Main.py based on
user input for the number of pennies and the number of seconds.