Python practice from December 7th, 2015
Reads in from a file .txt file containing a list of names, alphabetizes
the names and prints the names to the command console in a list style.
Takes in user input for the theoretical price of a pizza along with its
radius and determines the price for one slice of the pizza.